Split out basic WebSocket server template
