+++ tests/echo.py
@@ -12,9 +12,9 @@ as taken from http://docs.python.org/dev/library/ssl.html#certificates
 
 import os, sys, select, optparse, logging
 sys.path.insert(0,os.path.join(os.path.dirname(__file__), ".."))
-from websockify.websocketserver import WebSocketServer, WebSocketRequestHandler
+from websockify.websockifyserver import WebSockifyServer, WebSockifyRequestHandler
 
-class WebSocketEcho(WebSocketRequestHandler):
+class WebSocketEcho(WebSockifyRequestHandler):
     """
     WebSockets server that echos back whatever is received from the
     client.  """
@@ -71,6 +71,6 @@ if __name__ == '__main__':
     logging.basicConfig(level=logging.INFO)
 
     opts.web = "."
-    server = WebSocketServer(WebSocketEcho, **opts.__dict__)
+    server = WebSockifyServer(WebSocketEcho, **opts.__dict__)
     server.start_server()
