+++ websockify/websocketproxy.py
@@ -17,7 +17,7 @@ except: from SocketServer import ForkingMixIn
 try:    from http.server import HTTPServer
 except: from BaseHTTPServer import HTTPServer
 import select
-from websockify import websocketserver
+from websockify import websockifyserver
 from websockify import auth_plugins as auth
 try:
     from urllib.parse import parse_qs, urlparse
@@ -25,7 +25,7 @@ except:
     from cgi import parse_qs
     from urlparse import urlparse
 
-class ProxyRequestHandler(websocketserver.WebSocketRequestHandler):
+class ProxyRequestHandler(websockifyserver.WebSockifyRequestHandler):
 
     buffer_size = 65536
 
@@ -88,7 +88,7 @@ Traffic Legend:
             msg += " (using SSL)"
         self.log_message(msg)
 
-        tsock = websocketserver.WebSocketServer.socket(self.server.target_host,
+        tsock = websockifyserver.WebSockifyServer.socket(self.server.target_host,
                                                        self.server.target_port,
                                                        connect=True,
                                                        use_ssl=self.server.ssl_target,
@@ -221,7 +221,7 @@ Traffic Legend:
                 cqueue.append(buf)
                 self.print_traffic("{")
 
-class WebSocketProxy(websocketserver.WebSocketServer):
+class WebSocketProxy(websockifyserver.WebSockifyServer):
     """
     Proxy traffic to and from a WebSockets client to a normal TCP
     socket server target.
@@ -274,7 +274,7 @@ class WebSocketProxy(websocketserver.WebSocketServer):
                 "REBIND_OLD_PORT": str(kwargs['listen_port']),
                 "REBIND_NEW_PORT": str(self.target_port)})
 
-        websocketserver.WebSocketServer.__init__(self, RequestHandlerClass, *args, **kwargs)
+        websockifyserver.WebSockifyServer.__init__(self, RequestHandlerClass, *args, **kwargs)
 
     def run_wrap_cmd(self):
         self.msg("Starting '%s'", " ".join(self.wrap_cmd))
@@ -469,7 +469,7 @@ def websockify_init():
         if len(args) > 2:
             parser.error("Too many arguments")
 
-    if not websocketserver.ssl and opts.ssl_target:
+    if not websockifyserver.ssl and opts.ssl_target:
         parser.error("SSL target requested and Python SSL module not loaded.");
 
     if opts.ssl_only and not os.path.exists(opts.cert):
